[FEAT] 센트리 sdk 적용
