+++ package.json
@@ -41,6 +41,7 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
+    "@sentry/node": "^7.57.0",
     "@slack/client": "^5.0.2",
     "@types/bcrypt": "^5.0.0",
     "@types/bcryptjs": "^2.4.2",
